# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Chronicle TSDB",
    "description": "Time-series database HTTP API",
    "version": "0.2.0",
    "license": "MIT",
    "contact": {
      "name": "Chronicle Contributors"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8086",
      "description": "Chronicle server"
    }
  ],
  "paths": {
    "/api/v1/admin/compact": {
      "post": {
        "operationId": "compactDB",
        "summary": "Trigger storage compaction",
        "tags": [
          "admin"
        ],
        "responses": {
          "200": {
            "description": "Compaction started"
          },
          "500": {
            "description": "Compaction failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/flush": {
      "post": {
        "operationId": "flushDB",
        "summary": "Flush write buffer to storage",
        "tags": [
          "admin"
        ],
        "responses": {
          "200": {
            "description": "Flush successful"
          },
          "500": {
            "description": "Flush failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/cql": {
      "post": {
        "operationId": "executeCQL",
        "summary": "Execute a CQL (Chronicle Query Language) query",
        "tags": [
          "query"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "query": {
                    "type": "string",
                    "description": "CQL query string"
                  }
                },
                "required": [
                  "query"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Query results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Feature disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cql/validate": {
      "post": {
        "operationId": "validateCQL",
        "summary": "Validate a CQL query without executing it",
        "tags": [
          "query"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "query": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Query is valid"
          },
          "400": {
            "description": "Query is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/flags/toggle": {
      "post": {
        "operationId": "toggleFeatureFlag",
        "summary": "Enable or disable a feature flag",
        "tags": [
          "features"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "enabled": {
                    "type": "boolean",
                    "description": "Desired state"
                  },
                  "name": {
                    "type": "string",
                    "description": "Feature flag name"
                  }
                },
                "required": [
                  "name",
                  "enabled"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Toggle successful"
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Feature is protected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "description": "Standardized error response.",
        "properties": {
          "code": {
            "type": "integer",
            "description": "HTTP status code"
          },
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "error_type": {
            "type": "string",
            "description": "Error category (e.g., 'timeout', 'validation')"
          },
          "request_id": {
            "type": "string",
            "description": "Request correlation ID"
          },
          "status": {
            "type": "string",
            "description": "Always 'error'",
            "example": "error"
          }
        },
        "required": [
          "status",
          "error",
          "code"
        ]
      },
      "Point": {
        "type": "object",
        "description": "A single time-series data point.",
        "properties": {
          "metric": {
            "type": "string",
            "description": "Metric name"
          },
          "tags": {
            "type": "object",
            "description": "Key-value tag pairs"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64",
            "description": "Unix timestamp in nanoseconds"
          },
          "value": {
            "type": "number",
            "format": "double",
            "description": "Data point value"
          }
        },
        "required": [
          "metric",
          "value",
          "timestamp"
        ]
      },
...
      "QueryResult": {
        "type": "object",
        "description": "Result of a time-series query.",
        "properties": {
          "points": {
            "type": "array",
            "description": "Resulting data points",
            "items": {
              "$ref": "#/components/schemas/Point"
            }
          }
        }
      }
    }
  }
}